ixdtf::parsers

Module datetime

Parsing for Date, DateTime, and MonthDay.

Structs

DateTimeRecord 🔒

A DateTime Parse Node that contains the date, time, and offset info.

Functions

check_date_validity 🔒

days_in_month 🔒

Utilty to return the days in month, returns None if month is invalid

in_leap_year 🔒

Utility that returns whether a year is a leap year.

is_valid_month_day 🔒

parse_annotated_date_time 🔒

This function handles parsing for AnnotatedDateTime, AnnotatedDateTimeTimeRequred, and TemporalInstantString. according to the requirements provided via Spec.

parse_annotated_month_day 🔒

Parses an AnnotatedMonthDay.

parse_annotated_year_month 🔒

Parse an annotated YearMonth

parse_date 🔒

Parses Date record.

parse_date_day 🔒

parse_date_month 🔒

parse_date_time 🔒

Parses a DateTime record.

parse_date_year 🔒

parse_month_day 🔒

Parses a DateSpecMonthDay

parse_year_month 🔒