ixdtf

Module records

The records that ixdtf’s contain the resulting values of parsing.

Structs

Annotation

A record of an annotation.

DateRecord

The record of a parsed date.

Fraction

A fraction value in nanoseconds or lower value.

FullPrecisionOffset

A full precision UTC offset represented by a MinutePrecisionOffset with seconds and an optional fractional seconds

IxdtfParseRecord

An IxdtfParseRecord is an intermediary record returned by IxdtfParser.

MinutePrecisionOffset

A minute preicision UTC offset

TimeRecord

Parsed Time info

TimeZoneAnnotation

A TimeZoneAnnotation that represents a parsed TimeZoneRecord and its critical flag.

Enums

Sign

The parsed sign value, representing whether its struct is positive or negative.

TimeZoneRecord

Parsed TimeZone data, which can be either a UTC Offset value or IANA Time Zone Name value.

UtcOffsetRecord

A UtcOffsetRecord that is either a minute precision or full precision UTC offset.

UtcOffsetRecordOrZ